EES-2544 use worker for csv download
