+++ src/explore-education-statistics-common/__mocks__/mock.worker.ts
@@ -1,0 +1,11 @@
+export default class MockWorker {
+  onmessage: () => void;
+
+  constructor() {
+    this.onmessage = () => {};
+  }
+
+  postMessage() {
+    this.onmessage();
+  }
+}
