+++ src/explore-education-statistics-common/package.json
@@ -76,11 +76,13 @@
     "react-modal": "^3.12.1",
     "react-test-renderer": "16.13.1",
     "recharts": "^1.8.5",
+    "regenerator-runtime": "0.13.9",
     "sanitize-html": "^2.0.0",
     "stylelint": "^13.13.1",
     "typescript": "^4.0.3",
     "use-immer": "^0.4.0",
     "uuid": "^8.3.1",
+    "worker-loader": "^2.0.0",
     "xlsx": "^0.16.8",
     "yup": "^0.29.3"
   },
@@ -134,7 +136,8 @@
     "moduleNameMapper": {
       "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy",
       "^@common/(.*)$": "<rootDir>/src/$1",
-      "^@common-test/(.*)$": "<rootDir>/test/$1"
+      "^@common-test/(.*)$": "<rootDir>/test/$1",
+      "\\.worker.js":"<rootDir>/__mocks__/mock.worker.ts"
     },
     "moduleFileExtensions": [
       "web.js",
