+++ src/explore-education-statistics-common/src/modules/table-tool/components/utils/__tests__/downloadCsvUtils.test.ts
@@ -1,9 +1,7 @@
 import {
-  appendColumnWidths,
-  appendFootnotes,
-  appendTitle,
+  downloadCsvFile,
   getCsvData,
-} from '@common/modules/table-tool/components/utils/downloadTableUtils';
+} from '@common/modules/table-tool/components/utils/downloadCsvUtils';
 import {
   CategoryFilter,
   Indicator,
@@ -11,188 +9,70 @@ import {
   TimePeriodFilter,
 } from '@common/modules/table-tool/types/filters';
 import { FullTableMeta } from '@common/modules/table-tool/types/fullTable';
-import { utils } from 'xlsx';
+import { waitFor } from '@testing-library/react';
+import { writeFile, WorkBook } from 'xlsx';
 
-describe('Download Table utils', () => {
-  describe('appendColumnWidths', () => {
-    test('sets column widths using the cells with the longest string content', () => {
-      const sheet = utils.aoa_to_sheet([
-        ['test', 'test', 'testtest'],
-        ['test', 'testtest', 'test'],
-        ['testtest', 'test', 'testtesttest'],
-      ]);
+jest.mock('xlsx', () => {
+  const { utils } = jest.requireActual('xlsx');
 
-      appendColumnWidths(sheet);
+  return {
+    writeFile: jest.fn(),
+    utils,
+  };
+});
 
-      expect(sheet['!cols']).toEqual([{ wch: 8 }, { wch: 8 }, { wch: 12 }]);
-    });
-
-    test('sets column widths using the cells with the longest number content', () => {
-      const sheet = utils.aoa_to_sheet([
-        ['test', 'test', 'testtest'],
-        ['test', 12345678, 'test'],
-        ['testtest', 'test', 'testtesttest'],
-      ]);
-
-      appendColumnWidths(sheet);
-
-      expect(sheet['!cols']).toEqual([{ wch: 8 }, { wch: 8 }, { wch: 12 }]);
-    });
-  });
-
-  describe('appendTitle', () => {
-    test('adds title to start of sheet', () => {
-      const sheet = utils.aoa_to_sheet([
-        ['test', 'test', 'test'],
-        ['test', 'test', 'test'],
-        ['test', 'test', 'test'],
-      ]);
-
-      expect(sheet['!ref']).toBe('A1:C3');
-
-      appendTitle(sheet, 'Test title');
-
-      expect(sheet['!ref']).toBe('A1:C5');
-
-      expect(sheet.A1.v).toBe('Test title');
-      expect(sheet.A2.v).toBe('');
-      expect(sheet.A3.v).toBe('test');
-    });
-
-    test('preserves empty existing cells', () => {
-      const sheet = utils.aoa_to_sheet([
-        ['', '', 'test'],
-        ['test', 'test', 'test'],
-        ['test', 'test', 'test'],
-      ]);
-
-      expect(sheet['!ref']).toBe('A1:C3');
-
-      appendTitle(sheet, 'Test title');
-
-      expect(sheet['!ref']).toBe('A1:C5');
-
-      expect(sheet.A1.v).toBe('Test title');
-      expect(sheet.A2.v).toBe('');
-      expect(sheet.A3.v).toBe('');
-      expect(sheet.B3.v).toBe('');
-    });
-  });
-
-  describe('appendFootnotes', () => {
-    test('adds single footnote to end of sheet', () => {
-      const sheet = utils.aoa_to_sheet([
-        ['test', 'test', 'test'],
-        ['test', 'test', 'test'],
-        ['test', 'test', 'test'],
-      ]);
-
-      expect(sheet['!ref']).toBe('A1:C3');
-
-      appendFootnotes(sheet, [
-        {
-          id: '1',
-          label: 'Test footnote 1',
-        },
-      ]);
-
-      expect(sheet['!ref']).toBe('A1:C5');
-      expect(sheet.A5).toEqual({
-        t: 's',
-        v: '(1) Test footnote 1',
-      });
-    });
-
-    test('adds multiple footnotes to end of sheet', () => {
-      const sheet = utils.aoa_to_sheet([
-        ['test', 'test', 'test'],
-        ['test', 'test', 'test'],
-        ['test', 'test', 'test'],
-      ]);
-
-      expect(sheet['!ref']).toBe('A1:C3');
-
-      appendFootnotes(sheet, [
-        {
-          id: '1',
-          label: 'Test footnote 1',
-        },
-        {
-          id: '2',
-          label: 'Test footnote 2',
-        },
-        {
-          id: '3',
-          label: 'Test footnote 3',
-        },
-      ]);
-
-      expect(sheet['!ref']).toBe('A1:C7');
-      expect(sheet.A5).toEqual({
-        t: 's',
-        v: '(1) Test footnote 1',
-      });
-      expect(sheet.A6).toEqual({
-        t: 's',
-        v: '(2) Test footnote 2',
-      });
-      expect(sheet.A7).toEqual({
-        t: 's',
-        v: '(3) Test footnote 3',
-      });
-    });
-  });
+describe('Download CSV utils', () => {
+  const basicSubjectMeta: FullTableMeta = {
+    geoJsonAvailable: false,
+    publicationName: '',
+    subjectName: '',
+    footnotes: [],
+    boundaryLevels: [],
+    filters: {
+      Characteristic: {
+        name: 'characteristic',
+        options: [
+          new CategoryFilter({
+            value: 'gender_female',
+            label: 'Female',
+            group: 'Gender',
+            category: 'Characteristic',
+          }),
+        ],
+      },
+    },
+    indicators: [
+      new Indicator({
+        label: 'Authorised absence rate',
+        value: 'authAbsRate',
+        unit: '%',
+        name: 'sess_authorised_percent',
+      }),
+      new Indicator({
+        label: 'Number of authorised absence sessions',
+        value: 'authAbsSess',
+        unit: '',
+        name: 'sess_authorised',
+      }),
+    ],
+    locations: [
+      new LocationFilter({
+        value: 'england',
+        label: 'England',
+        level: 'country',
+      }),
+    ],
+    timePeriodRange: [
+      new TimePeriodFilter({
+        code: 'AY',
+        year: 2015,
+        label: '2015/16',
+        order: 0,
+      }),
+    ],
+  };
 
   describe('getCsvData', () => {
-    const basicSubjectMeta: FullTableMeta = {
-      geoJsonAvailable: false,
-      publicationName: '',
-      subjectName: '',
-      footnotes: [],
-      boundaryLevels: [],
-      filters: {
-        Characteristic: {
-          name: 'characteristic',
-          options: [
-            new CategoryFilter({
-              value: 'gender_female',
-              label: 'Female',
-              group: 'Gender',
-              category: 'Characteristic',
-            }),
-          ],
-        },
-      },
-      indicators: [
-        new Indicator({
-          label: 'Authorised absence rate',
-          value: 'authAbsRate',
-          unit: '%',
-          name: 'sess_authorised_percent',
-        }),
-        new Indicator({
-          label: 'Number of authorised absence sessions',
-          value: 'authAbsSess',
-          unit: '',
-          name: 'sess_authorised',
-        }),
-      ],
-      locations: [
-        new LocationFilter({
-          value: 'england',
-          label: 'England',
-          level: 'country',
-        }),
-      ],
-      timePeriodRange: [
-        new TimePeriodFilter({
-          code: 'AY',
-          year: 2015,
-          label: '2015/16',
-          order: 0,
-        }),
-      ],
-    };
     test('contains full set of data', async () => {
       const data = getCsvData({
         subjectMeta: {
@@ -543,6 +423,33 @@ describe('Download Table utils', () => {
       expect(data[1][6]).toBe('x');
 
       expect(data).toMatchSnapshot();
+    });
+  });
+
+  describe('downloadCsvFile', () => {
+    test('creates and downloads the csv file', async () => {
+      const data = getCsvData({
+        subjectMeta: basicSubjectMeta,
+        results: [],
+      });
+
+      downloadCsvFile(data, 'The file name');
+
+      const mockedWriteFile = writeFile as jest.Mock;
+      await waitFor(() => {
+        expect(mockedWriteFile).toHaveBeenCalledTimes(1);
+
+        const workbook = mockedWriteFile.mock.calls[0][0] as WorkBook;
+
+        expect(workbook.Sheets.Sheet1.A1.v).toBe('location');
+        expect(workbook.Sheets.Sheet1.B1.v).toBe('location_code');
+        expect(workbook.Sheets.Sheet1.C1.v).toBe('geographic_level');
+        expect(workbook.Sheets.Sheet1.D1.v).toBe('time_period');
+        expect(workbook.Sheets.Sheet1.E1.v).toBe('characteristic');
+        expect(workbook.Sheets.Sheet1.F1.v).toBe('sess_authorised_percent');
+
+        expect(mockedWriteFile.mock.calls[0][1]).toBe('The file name.csv');
+      });
     });
   });
 });
