+++ src/explore-education-statistics-frontend/src/types/worker-loader.d.ts
@@ -1,0 +1,7 @@
+declare module 'worker-loader?name=static/[hash].worker.js!*' {
+  class WebpackWorker extends Worker {
+    public constructor();
+  }
+
+  export default WebpackWorker;
+}
